requirements.txt
pandas
numpy
matplotlib
plotly
geopandas
shapely
pyproj
scikit-learn
xgboost
lightgbm
joblib
pyspark
dvc
gitpython
pyyaml
